pyproject.toml
[project]
name = "python-project-83"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
    "validators>=0.35.0",
]
[project.scripts]
page-analyzer = "page_analyzer.app:app"
[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
